+++ packages/server/lib/hooks/connection/internal-nango.ts
@@ -1,5 +1,6 @@
-import { ProxyRequest, connectionService, getProxyConfiguration } from '@nangohq/shared';
+import { ProxyRequest, configService, connectionService, getProxyConfiguration } from '@nangohq/shared';
 
+import type { Config } from '@nangohq/shared';
 import type { ConnectionConfig, DBConnectionDecrypted, InternalProxyConfiguration, Provider, UserProvidedProxyConfiguration } from '@nangohq/types';
 import type { AxiosError, AxiosResponse } from 'axios';
 
@@ -8,6 +9,7 @@ export interface InternalNango {
     proxy: <T = any>({ method, endpoint, data, headers, params, baseUrlOverride }: UserProvidedProxyConfiguration) => Promise<AxiosResponse<T> | AxiosError>;
     updateConnectionConfig: (config: ConnectionConfig) => Promise<ConnectionConfig>;
     unsetConnectionConfigAttributes: (...keys: string[]) => Promise<ConnectionConfig>;
+    getIntegration: () => Promise<Config | null>;
 }
 
 export function getInternalNango(connection: DBConnectionDecrypted, providerName: string): InternalNango {
@@ -49,6 +51,9 @@ export function getInternalNango(connection: DBConnectionDecrypted, providerName
         },
         unsetConnectionConfigAttributes: (...keys: string[]) => {
             return connectionService.unsetConnectionConfigAttributes(connection, keys);
+        },
+        getIntegration: async () => {
+            return await configService.getProviderConfig(connection.provider_config_key, connection.environment_id);
         }
     };
 }