+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v0.69.2] - 2025-10-06
+
+### Added
+
+- *(integrations)* Add support for folk (#4801) by @hassan254-prog
+
+### Changed
+
+- Clean logs (#4812) by @TBonnin
+
+### Fixed
+
+- *(kvstore)* Set defaults on getRedisClient (#4811) by @rossmcewan
+- Modify task processor tests to use wildcard groupKeyPattern (#4800) by @TBonnin
+- Reverting recent redis change (#4815) by @TBonnin
+- *(connect-ui)* Pass oauth2-cc connection config defaults (#4813) by @hassan254-prog
+- *(cli-tests)* Update path with changed api (#4817) by @khaliqgant
+
 ## [v0.69.1] - 2025-10-03
 
 ### Added
@@ -5055,6 +5073,7 @@ All notable changes to this project will be documented in this file.
 - Update Zendesk Chat configuration file (#161) by @Frenchcooc
 - APIs using client_credentials as grant type (#165) by @Frenchcooc
 
+[v0.69.2]: https://github.com/NangoHQ/nango/compare/v0.69.1..v0.69.2
 [v0.69.1]: https://github.com/NangoHQ/nango/compare/v0.69.0..v0.69.1
 [v0.69.0]: https://github.com/NangoHQ/nango/compare/v0.68.0..v0.69.0
 [v0.68.0]: https://github.com/NangoHQ/nango/compare/v0.67.8..v0.68.0